requirements.txt
python-telegram-bot
python-dotenv
aiohttp
bcrypt
zhipuai
httpx
docker
croniter
google-generativeai
